backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
pillow
opencv-python-headless
easyocr
sentence-transformers
faiss-cpu
reportlab
qrcode[pil]
beautifulsoup4
requests
feedparser
slowapi
python-dotenv
httpx
torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
pillow>=10.0.0
scikit-learn>=1.3.0
sentencepiece
